client/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyio==4.4.0
    # via httpx
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpx[http2]==0.27.0
    # via microsoft-kiota-http
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
importlib-metadata==8.0.0
    # via opentelemetry-api
microsoft-kiota-abstractions==1.3.3
    # via
    #   -r requirements.in
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
microsoft-kiota-http==1.3.2
    # via -r requirements.in
microsoft-kiota-serialization-form==0.1.0
    # via -r requirements.in
microsoft-kiota-serialization-json==1.3.0
    # via -r requirements.in
microsoft-kiota-serialization-multipart==0.1.0
    # via -r requirements.in
microsoft-kiota-serialization-text==1.0.0
    # via -r requirements.in
opentelemetry-api==1.26.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.26.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.47b0
    # via opentelemetry-sdk
pendulum==3.0.0
    # via
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
python-dateutil==2.9.0.post0
    # via
    #   microsoft-kiota-serialization-text
    #   pendulum
    #   time-machine
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
std-uritemplate==1.0.5
    # via microsoft-kiota-abstractions
time-machine==2.15.0
    # via pendulum
typing-extensions==4.12.2
    # via opentelemetry-sdk
tzdata==2024.1
    # via pendulum
wrapt==1.16.0
    # via deprecated
zipp==3.19.2
    # via importlib-metadata
